Moved visitor pattern classes to a visit sub-directory for both includes and sources.
