+++ CMakeLists.txt
@@ -18,7 +18,7 @@ include_directories(include)
 
 # All cmmcore source files
 set(SOURCE_FILES src/AddressOfNode.cpp src/ArgNode.cpp src/BinOpNode.cpp src/BlockNode.cpp
-    src/CompilationUnitNode.cpp src/DerefNode.cpp src/Dump.cpp
+    src/CompilationUnitNode.cpp src/DerefNode.cpp
     src/ExpressionNode.cpp src/ExpressionStatementNode.cpp src/FunctionDeclarationStatementNode.cpp
     src/Frame.cpp src/FunctionCallNode.cpp src/FunctionDefinitionStatementNode.cpp src/IfElseStatementNode.cpp src/Lexer.cpp
     src/LitteralNode.cpp src/Location.cpp src/Node.cpp src/ParameterNode.cpp src/ParenExpressionNode.cpp
@@ -26,7 +26,8 @@ set(SOURCE_FILES src/AddressOfNode.cpp src/ArgNode.cpp src/BinOpNode.cpp src/Blo
     src/Reporter.cpp src/ReturnStatementNode.cpp src/Snapshot.cpp src/StatementNode.cpp src/StringView.cpp
     src/Scope.cpp src/Token.cpp
     src/TranslationUnitNode.cpp src/Types.cpp src/TypeNode.cpp
-    src/VariableNode.cpp src/VariableDeclarationStatementNode.cpp src/Visitor.cpp src/WhileStatementNode.cpp)
+    src/VariableNode.cpp src/VariableDeclarationStatementNode.cpp src/WhileStatementNode.cpp
+    src/visit/Dump.cpp src/visit/Visitor.cpp)
 
 # Require threading library (Linux/Unix this is pthreads).
 set(THREADS_PREFER_PTHREAD_FLAG ON)
