+++ include/cmm/visit/Dump.h
@@ -13,7 +13,7 @@
 // Our includes
 #include <cmm/Types.h>
 #include <cmm/NodeListFwd.h>
-#include <cmm/Visitor.h>
+#include <cmm/visit/Visitor.h>
 
 namespace cmm
 {
