+++ src/Main.cpp
@@ -1,7 +1,7 @@
 // Our includes
-#include <cmm/Dump.h>
 #include <cmm/NodeList.h>
 #include <cmm/Parser.h>
+#include <cmm/visit/Dump.h>
 
 // std includes
 #include <iostream>
